server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file uploads

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
# chromadb
pinecone

# Embeddings
#sentence-transformers
#langchain-google-generative-ai
#langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging (optional but recommended)
loguru

pinecone-client
streamlit
requests

#  uv pip install -r requirements.txt
#uvicorn main:app --reload
#pip install -r server/requirements.txt
#pip install -r client/requirements.txt